refactor(renderer): move _new_design profile + settings to self/
Second slice of dissolving _new_design/. The agent's own surfaces move to
a new top-level self/:
- self/profile/  — moss-create/edit/detail/prompt-profile, profile-settings
- self/settings/ — moss-settings, language/notification/danger-zone-settings

Pure file moves + import-path updates; no behaviour change.
