+++ src/renderer/src/elements/_new_design/group-settings/my-profile-settings.ts
@@ -13,7 +13,7 @@ import { MossStore } from '../../../moss-store.js';
 import { groupStoreContext } from '../../../groups/context.js';
 import { GroupStore } from '../../../groups/group-store.js';
 import { mossStyles } from '../../../shared-styles.js';
-import { MossEditProfile } from '../profile/moss-edit-profile.js';
+import { MossEditProfile } from '../../../self/profile/moss-edit-profile.js';
 import { Profile } from '@holochain-open-dev/profiles';
 import { EntryRecord } from '@holochain-open-dev/utils';
